requirements.txt
#required libraries
aiohttp==3.6.2
fastai==1.0.60
Pillow==7.1.1
python-multipart==0.0.5
starlette==0.13.3
torch==1.5.0
torchvision==0.6.0
uvicorn==0.11.3
Flask == 1.1.2
numpy == 1.19.0
requests == 2.24.0
opencv-python == 4.4.0.42
tensorflow==2.3.0
h5py==2.10.0
Keras==2.1.5
google-cloud-storage==1.31.2
